requirements.txt
python-dotenv~=0.20.0
neuralintents
youtube_dl
PyNaCl
requests~=2.28.1
gptj
py-cord
jishaku
discord-pretty-help
motor
psutil
nest_asyncio
dnspython
pip-upgrader
wikipedia~=1.4.0
humanize~=4.3.0
PyQRCode~=1.2.1
DateTime~=4.5
aiohttp~=3.7.4.post0